<Added Response Entity>

+++ src/main/java/com/food/Restaurant/controller/RestaurantController.java
@@ -4,6 +4,8 @@ import com.food.Restaurant.entity.Restaurant;
 import com.food.Restaurant.service.RestaurantService;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,34 +18,53 @@ public class RestaurantController {
     private RestaurantService restaurantService;
 
     @GetMapping
-    public List<Restaurant> getAll() {
-        return restaurantService.getAllRestaurants();
+    public ResponseEntity<?> getAll() {
+        List<Restaurant> allEntries = restaurantService.getAllRestaurants();
+        if (allEntries != null && !allEntries.isEmpty()) {
+            return new ResponseEntity<>(allEntries, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("No restaurants to show in the collection", HttpStatus.NOT_FOUND);
+        }
     }
 
     @PostMapping
-    public String saveRestaurant(@RequestBody Restaurant restaurantBody) {
-        double ratings = restaurantBody.getRatings();
-        if (ratings > 5.0 || ratings < 0.0) {
-            return "Ratings should be in the range 0.0-5.0";
-        } else {
-            restaurantService.saveRestaurant(restaurantBody);
-            return "ID of New Restaurant: " + restaurantBody.getId();
+    public ResponseEntity<?> saveRestaurant(@RequestBody Restaurant restaurantBody) {
+        try {
+            double ratings = restaurantBody.getRatings();
+            if (ratings > 5.0 || ratings < 0.0) {
+                return new ResponseEntity<>("Ratings should be in the range 0.0-5.0", HttpStatus.BAD_REQUEST);
+            } else {
+                restaurantService.saveRestaurant(restaurantBody);
+                return new ResponseEntity<>("ID of New Restaurant: " + restaurantBody.getId(), HttpStatus.CREATED);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
 
     @GetMapping("/id/{myID}")
-    public Restaurant getRestaurantById(@PathVariable ObjectId myID) {
-        return restaurantService.getRestaurantById(myID).orElse(null);
+    public ResponseEntity<?> getRestaurantById(@PathVariable ObjectId myID) {
+        Restaurant restaurant = restaurantService.getRestaurantById(myID).orElse(null);
+        if (restaurant != null) {
+            return new ResponseEntity<>(restaurant, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("ID not found", HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/id/{myID}")
-    public String deleteRestaurantById(@PathVariable ObjectId myID) {
-        restaurantService.deleteRestaurantById(myID);
-        return "ID of Deleted Restaurant: " + myID;
+    public ResponseEntity<?> deleteRestaurantById(@PathVariable ObjectId myID) {
+        Restaurant restaurant = restaurantService.getRestaurantById(myID).orElse(null);
+        if (restaurant != null) {
+            restaurantService.deleteRestaurantById(myID);
+            return new ResponseEntity<>("ID of Deleted Restaurant: " + myID, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("ID not found", HttpStatus.NOT_FOUND);
+        }
     }
 
     @PutMapping("/id/{myID}")
-    public String updateRestaurantById(@PathVariable ObjectId myID, @RequestBody Restaurant newRestaurant) {
+    public ResponseEntity<?> updateRestaurantById(@PathVariable ObjectId myID, @RequestBody Restaurant newRestaurant) {
         Restaurant oldRestaurant = restaurantService.getRestaurantById(myID).orElse(null);
         if (oldRestaurant != null) {
             //IF newRestaurant's <parameter> is blank set the oldRestaurant's <parameter> as itself
@@ -54,9 +75,9 @@ public class RestaurantController {
             oldRestaurant.setLocation(newRestaurant.getLocation().isBlank() ? oldRestaurant.getLocation() : newRestaurant.getLocation());
             oldRestaurant.setRatings(newRestaurant.getRatings() > 5.0 || newRestaurant.getRatings() < 0.0 ? oldRestaurant.getRatings() : newRestaurant.getRatings());
             restaurantService.saveRestaurant(oldRestaurant);
-            return "ID of Updated Restaurant: " + myID;
+            return new ResponseEntity<>("ID of Updated Restaurant: " + myID, HttpStatus.OK);
         } else {
-            return "ID " + myID + " not found";
+            return new ResponseEntity<>("ID not found", HttpStatus.NOT_FOUND);
         }
     }
 }
